Create Program16_5.java
The given Java program consists of two classes: Logic and Program5. The Logic class has a method countDigits that calculates the number of digits in a given integer num. It uses a while loop to divide the number by 10 repeatedly, incrementing a counter count until the number becomes zero. For the special case when the number is 0, the count is set to 1. The Program5 class contains the main method, where an object of Logic is created and countDigits is called with the number 7865. The program then prints the total number of digits in the given number.