Create Program18_1.java
The given Java program consists of two classes: Logic and Program1. The Logic class contains a method checkPrime that determines whether a given integer num is a prime number. The method first checks if the number is less than or equal to 1, in which case it is not prime. For numbers greater than 1, a boolean variable isPrime is initialized to true, and a for loop iterates from 2 to num/2 to check for divisors. If any number divides num evenly, isPrime is set to false and the loop breaks. After the loop, the program prints whether the number is prime based on the value of isPrime. The Program1 class contains the main method, which serves as the program's entry point. In main, an object of the Logic class is created, and checkPrime is called with the number 11. The program then outputs whether the given number is prime, demonstrating the use of loops, conditionals, and basic object-oriented programming in Java.