Create Program18_2.java
The given Java program consists of two classes: Logic and Program2. The Logic class has a method printEvenNumbers that prints all even numbers from 1 up to a given number n. It uses a for loop to iterate through numbers from 1 to n and checks each number with the modulus operator % to determine if it is even. The Program2 class contains the main method, where an object of Logic is created, and printEvenNumbers is called with the value 20. The program then displays all even numbers up to 20.