Create Program18_3.java
The given Java program consists of two classes: Logic and Program3. The Logic class contains a method printOddNumbers that prints all odd numbers from 1 up to a given number n. It uses a for loop to iterate through numbers from 1 to n and checks each number with the modulus operator % to determine if it is odd. The Program3 class contains the main method, where an object of Logic is created, and printOddNumbers is called with the value 20. The program then displays all odd numbers up to 20.