Create Program18_5.java
The given Java program consists of two classes: Logic and Program. The Logic class has a method checkSign that checks whether a given number num is positive, negative, or zero using conditional statements. The Program class contains the main method, where an object of Logic is created, and checkSign is called with the value -8. The program then prints the sign of the number.