Create Program19_3.java
This Java program checks whether a given number is divisible by both 5 and 11. The Logic class has a method checkDivisible(int num) that uses the modulus operator % and an if-else statement to determine divisibility. If the number is divisible by both 5 and 11, it prints a positive message; otherwise, it indicates that it is not. In the Program3 class, an object of Logic is created, and the method is called to check the number 55. The program demonstrates conditional checks and logical operators in Java.