Create Program20_1.java
This Java program calculates the sum of all even numbers up to a given number N. The Logic class contains the method sumEvenNumbers(int n), which uses a for loop starting from 2 and increments by 2 to add all even numbers up to n. The sum is then printed. In the Program1 class, an object of Logic is created, and the method is called to find the sum of even numbers up to 10. The program demonstrates loops, arithmetic operations, and accumulation in Java.