Create Program21_3.java
This Java program displays all the factors of a given number. The Logic class defines the method displayFactors(int num), which uses a for loop to check each number from 1 to num. If the current number divides num exactly (remainder 0), it is printed as a factor. In the Program3 class, an object of Logic is created, and the method is called to display the factors of 12. The program demonstrates loops, conditional checks, and factor determination in Java.