Create Program21_4.java
This Java program counts the total number of factors of a given number. The Logic class includes the method countFactors(int num), which uses a for loop to check every number from 1 to num. If a number divides num completely (remainder 0), it is counted as a factor. After checking all numbers, the total count of factors is displayed. In the Program4 class, an object of Logic is created, and the method is called to count the factors of 20. The program demonstrates loops, conditional statements, and factor counting in Java.